HYD-4520 Test failure in test_create_single_filesystem_add_mdt test
The test_create_single_filesystem_add_mdt is causing lustre to raise an
error with a stderr of "error on LL_IOC_LMV_SETSTRIPE '/mnt/testfs/mdt1'
(3): No such file or directory
error: mkdir: create stripe dir '/mnt/testfs/mdt1'"

This error is causing issues with landing follow on patches and so this
patch adds a skip around the error whilst we work through the root cause.
There is always debate about removing vs. commenting out but in this case
it seems to me that commenting out for a blocking issue is probably the best
course of action.

  - Add skip around the test_create_single_filesystem_add_mdt
  - Correct the typo in the name of the test_create_single_filesystem_add_mdt
    test (I just couldn't not remove the trailing d on singled

Change-Id: I54f289a2471cc154995c65ef2187d319102cda03
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15351
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>