HYD-4533 Automate Angular DI compilation protection
The compilation step would break the Angular Dependency
Injector (the identifiers would be skewed by the compiler)
if we did not use a wrapper block and define our function
arguments / dependencies twice; once for the function as
args, and once as a list of dependencies as strings.

Add automation to fix the problem.
Had to create a github ticket to fix the dependency issue
we were seeing in Jenkins.
https://github.com/olov/ng-annotate/issues/178

- package.json  Added gulp-ng-annotate lib.
- ziplock.json  Added gulp-ng-annotate lib.
- gulpfile.js  Integrated gulp-ng-annotate.

chroma-externals: 15363/9

Change-Id: I4021037bca04bfe12da3cbb4e50ff329952ebfd5
Reviewed-on: http://review.whamcloud.com/15364
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>