HYD-5539 Update to SLES11sp4 and add SLES12 client
EE 3.0 is intended to include both the SLES11sp4 server and
client, as well as a SLES12 client. These changes were done in
the past on prior branches mixed with other changes, but didn't
make it back to master.

This change pulls the necessary changes in to master to get the
SLES packaging up-to-date for 3.0.

  - Add new makefiles for sles12 in chroma-dependencies and chroma-bundles:
      - chroma-bundles/lustre-client-sles12/Makefile
      - chroma-dependencies/lustre-client-sles12/Makefile
  - Modify sles-client and sles-server Makefiles to refer to sles11sp4
    instead of sles11sp3.

Change-Id: I277a89575f16b4c3a02d1a11e4ff1e1b0136a33d
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/18650
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>