HYD-421 Generalise alerts and events
Alerts historically have been a mixture of 3 things. Commands,
Alerts and Events. This has made a general query of the sum
Notifications (the union of all 3) expensive in terms of db
queries. It also cause problems because the three record types
had different fields and so some awful fixing up of data was
carried out to make things appear good.  This mixed field types
also greatly restricted the filtering that could be applied via
the reset interface.

Worse still because of the way mti (multi table inheritance) works
in Django the number of queries to return an Alert or Event was
basically 1 query to fetch the initial records plus one addition
query for each record returned.

The solution is to scrap Events, and don't include Commands in
the Alerts (a later patch creates Command Alerts for commands). Then
we abandon the rubbish mti table inheritance and go for a less
flexible but infinately fast and better and easier to use sparse
table solution where only one table exists for all alerts. The
alerts are then created as they are queried removing the need to
ever downcast anything again.

old world
AlertState.objects.all() gives

[AlertState, AlertState, AlertState]

to find out what these actualy are you need to downcast each on
and each downcast is a query - and the original list gives debug
hell.

new world
AlertState.objects.all() gives

[HostOfflineAlert, HostRebootEvent, PacemakerStoppedAlert]

no need to ever downcast, no confusion, faster and easier to use.

There is some serious work to make this possible, and we may well expand
it in the future to use for things like ManagedTarget etc.

 - Create the SparseModel class to provide a better inheritance model.
 - Updated all alerts/events to use the new AlertStateBase class
 - Removed Command from the alerts
 - Deleted the Notifications endpoint
 - Deleted the Alerts endpoint

Change-Id: I633382fe75e43efd781d2a2f2795d08d3dc4e09a
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/16453
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>