HYD-4555 Add bindMethod to FP
We want an abstraction to create free floating functions
from object methods.  Furthermore, we usually need to
invoke the new function right away. This calls for a
second method on FP.

Create a new method on the FP library object that takes a
method name as a string and uses native bind to return a
new function that is not bound to an object.
Create a second new method that invokes this new method.

Added fp.bindMethod() and fp.invokeMethod.
Added the unit tests for fp.bindMethod() and fp.invokeMethod().

Change-Id: Iae2c46972c1987ab69565acfdb8b4c24f9d49e0b
Signed-off-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-on: http://review.whamcloud.com/15402
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>