HYD-4566 chroma-director is not used and should be removed.
chroma-director is a piece of AWS test development that
is not used and has not been updated since at least b2_1.

It's presence does have overhead and so it should be removed. If 
anyone wants to use its contents again in the future then they
can retrieve it from git.

  - remove the complete chroma-director directory

Change-Id: Ic5604d0c82e46728280a2b0e36980fa0d608a2cb
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15460
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>