HYD-4567 Make test update process use a single command.
The updates test uses 1 command per server when doing the
updates rather than the 1 command with many Jobs that the
GUI uses. Whilst both are valid ways of carrying out the
update it makes more sense to use the same method as the GUI.

Nice simplication of code to send a list of update jobs rather
than create a list of commands each with 1 update job. This
means that only 1 single command needs to be monitored for the
response.  Also added a little bit of extra debug into the
wait_for_command job which can make obscure issue debug easier.

 - Updated the Update test to use a single command with multiple
   jobs.

Change-Id: I528634abee7205da61d6ec7e9f24b46c4c9ffa26
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15461
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>