HYD-4568 refactor fp.and and fp.or
Currently fp.and() and fp.or() are variadic.
Recently, I ran into an issue that would
have been easier to reason about if these
methods took an array arg, instead of being
variadic.
Taking an array arg is how the similar methods
in fp work anyway.

Refactor them both and update the tests.

 - Updated the src and tests.

Change-Id: Ie71fd881a73c58682690c7e3c3c356fafa94ec83
Signed-off-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-on: http://review.whamcloud.com/15466
Tested-by: Jenkins
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>