HYD-4577 Add router to socket-worker
We would like to have data munging be off the main thread in
socket-worker. We need a nice way to have routable functions
to isolate data munging for specific paths.

Integrate the router module into socket-worker and
wire up a wildcard route. This will be equivalent
to what's in socket-worker today. Later on
we can add new routes as needed.

  - Add router to socket-worker.
  - Add wildcard route.
  - Add tests.

chroma-externals: 15488/1

Change-Id: Iaba0134b6cda45b35749af2a827452d002a59441
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/15489
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>