HYD-4451 Create a local-apply module
Currently, Angular.js has something called a $scope. A $scope is a tree
structure that closely mirrors a DOM tree. Whenever a change occurs,
Angular.js performs something called a digest cycle. A digest cycle
traverses the tree structure from the root and calls any watch functions
registered to each node. These watch functions tell Angular.js if
something has changed in the tree, which fires a cooresponding listener
function.

This works well but can become quite expensive as the number of watchers
in an application increases, leading to extra CPU time calculating
watches / jank.
Create a local-apply method. The local-apply method is responsible for
digesting a sub-tree starting at the node passed to it. This is much
more performant and scales better, as we are only computing changes for
the view we wish to update, not the entire application.

  - Create an extend-scope-module.
  - Create a local-apply method.
  - Create tests.

Change-Id: I950235f76d2407ad28943933f7e5b2e9f3079053
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/15503
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>