HYD-4435 Create block split-button mixin
In the server detail page we are using full width buttons
for a better mobile experience. Some of these buttons are
split buttons. Bootstrap does not support full-width
split buttons, so we need to create a mixin.

Create a mixin that we can call on a full-width
button to style it nicely.

  - Create a split-block-button mixin.

Change-Id: I860185423ff2d2f5d6d764777e8bd7e922b0a41a
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/15505
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>