HYD-4433 Implement Big Differ module
Our GUI is part of a distributed system. When the user is performing
actions that have the potential to deviate from known state without a
save (configure LNet, configure volumes, configure power control) We
should display the difference visually to the user and indicate how
their state is different from a known state.

We already do this in the LNet dialog, and somewhat on the volumes
screen. This ticket is to abstract the diffing logic to a reusable
component big-differ that will perform the diff calculations.

big-differ will take an initial collection, remote collection and a
local record, (or ax, bx and c to be more abstract) and return a change
item that tells if the local deviates from remote or initial.

  - Create big-differ-module.
  - Create big-differ tools.
  - Create big-differ tests.

Change-Id: I7d4c1b47b276d1e0c19f3a2760333633ccac2606
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/15506
Tested-by: Jenkins
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>