HYD-4585 Add "fastdom" 0.8.6 dependency to bower
In an effort to reduce DOM thrashing we've decided to add a library
called "fastdom" that should help reduce thrashing and ultimately
increase the speed of the gui.

This patch will update the bower.json file to include fastdom 0.8.6 as a
dependency and will also install the fastdom files in the
bower_components/fastdom directory.

- Updated bower.json to include fastdom 0.8.6
- Installed fastdom files in bower_components/fastdom directory

Change-Id: I70657272acbf4bb717bd38c13f63c32e4007d40b
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/15510
Tested-by: Jenkins
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>