HYD-4599 Update get-client-lib.js to start the primusPort on 8890 instead of 8889 as it is interfering with another process
After starting the manager I attempted to run "gulp build" to build out
my latest changes locally. To my surprise, I was receiving an EADDRESS
in use error. This can only mean that another process is using the port
the gulp is attempting to use. To fix this, a different port must be
used.

I'm updating the port in realtime/get-client-lib.js to 8890 instead of
8889. I do not see any instances of our services using this port so it
should be safe. I've tested this locally and it works fine.

- Updated get-client-lib to use port 8890

Change-Id: Ia183acd04d83fbf2669a5e26141195a81bc931b5
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/15514
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>