HYD-4602 Remove timecop dependency from ui
We no longer need the timecop dependency. The purpose of this patch is
to remove the dependency from the UI.

In this patch I will remove any timecop associated files from the
bower_components directory. The bower.json does not have timecop
specified so it does not need to be updated.

- Removed associated timecop source files from bower_components
- Updated read-write-heat-map-controller such that it creates a mock on
  Date and uses that mock instead of timecop to complete its test.

Change-Id: I41dddefa088c0ae202b46c3161d7b30a9e0bc477
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/15529
Tested-by: Jenkins
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>