HYD-4610 Add ability to capture selected test exceptions hold the system and request help.
Create a fetch_help function that will hold the node as is until
someone logs in an investigates.

The routine should allow the test to be allowed to continue by deleting
a file created and used as a key. The file will be called /tmp/waiting_help.
The routine should allow some validation of the exception by the caller
so that some exceptions can not be blocked.

 - Add _fetch_help function to the ApiTestCaseWithTestReset class
   this is called with the code to execute, list or people to email
   with message plus the option to have a callback that means the
   exception doesn't fetch help.

Change-Id: If8e80d6347317460c1b5f113ab2f91420827d58f
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15544
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>