HYD-4541 Update copyright message to include 2015
The copyright message in the source is outdated and should
include 2015. We need to update the source to require the
copyright on product source files and change current
copyrights to 2015.

When implementing this ticket I thought about whether only files
that changed in 2015 should have the copyright updated or all
files should have copyright updated.  I decided that only
files with a change in 2015, however minor, should have the copyright
updated and so did a global replace on every file to update
the copyright.

I did not correct other validation issues found in the code durring
commit as this was beyond this ticket. I did however update the
boilerplate.py script to accept the zero length files do not require
a copyright message.

  - All copyright messages updated.
  - boilerplate.py update to accept zero length files do not require
    a copyright message.

Change-Id: Ife225071e1233101c1a6329aa9a89290d82bd44c
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15379
Reviewed-by: Joe Grund <joe.grund@intel.com>
Tested-by: Jenkins
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>