HYD-4616 Convert relative paths in the cluster setup file to use the $MANAGER_DIR environment variable
The cluster setup file previously used relative paths to define module
locations. We've since added a new environment variable to represent the
chroma_manager directory. We should use this environment variable
instead of relative paths as it will be easier to read.

The patch will simply update the cluster_setup file to use the
$MANAGER_DIR environment variable instead of relative paths. No
additional logic will be changed.

  - Updated cluster_setup to use environment variables instead of
    relative paths

Change-Id: I59717e87fb2accdffae9b75a26ddc5b85bf95430
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/15587
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Tested-by: Joe Grund <joe.grund@intel.com>