HYD-4541 Update copyright message to include 2015
The copyright message in the source is outdated and should
include 2015. We need to update the source to require the
copyright on product source files and change current
copyrights to 2015.

This patch just updates the checks but does not update the files
this will make updating the files without a load of rebase issues
possible later on.

The boilerplate.py script is updated to accept the zero length
files do not require a copyright message.

  - boilerplate.py update to accept zero length files do not require
    a copyright message.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: Idbeec0626ab05c8106dd69f2d3663d67af9cc2b7
Reviewed-on: http://review.whamcloud.com/15588
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>