HYD-4629 chroma-diagnostics often hangs stalling the whole test queue.
chroma-diagnostics often hangs stalling the whole test queue, this could
be because of a number of things but fundamentally we must time it out
so that the queue does not stall.

Generally we should be very careful of calling external processes without
some timeout mechanism - so this ticket creates a function that can
be called in other places as well.

 - Add safe_shell module with run that has a timeout mechanism
 - Change chroma_log_collector.py to use that run instead of Popen

Change-Id: I5d9e450bb91d38bf941e1e95b152846ecd9addde
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15602
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
(cherry picked from commit f65de040be575ed30328cf433045dcb69c262244)
Reviewed-on: http://review.whamcloud.com/16002