HYD-4628 Number of client connections misreported when multiple MDT's exist
The number of clients reported by the filesystem endpoint is the
total connections to MDT's. As a client connects to each MDT in a
multi MDT system the number of connections in multiplied by the
number of MDTs. For example a system with 1 client and 3 MDTs
would report 3 clients because of the 3 connections.

The solution is to divide the total number of client connections
by the number of MDTs to give a true client connection count.

 - Update filesystem.py to deliver the number of clients not the
   number of client connections.

Change-Id: I9e0649e5db995764d7bdbbbb0713eeba73b90969
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15618
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>