HYD-4633 Create a library to handle various types of merging of objects
Previously we used the merge function in lo-dash to accomplish the
merging of objects. However, lo-dash is bloated and contains a lot of
extra code that is not needed. Since we are moving away from lo-dash, we
need a new module to compliment the fp library such that we have the
ability to merge objects, create shallow clones, etc. This functionality
is separate from the fp library but should instead compliment it and be
used in a more generic sense.

The purpose of this patch is to create the new library. It will exist
under:
chroma-manager/ui-modules/isomorphic/obj

This patch will add the following functionality to the obj library:
merge clone (shallow version)

  - cluster_setup: Added obj section
  - run_tests: Added obj section
  - test_dependencies: Added obj section
  - index: New obj library containing both "merge" and "clone"
  - jasmine: The test config
  - package: The npm package config
  - test/boot: The jasmine boot file for testing
  - test/tests: The jasmine tests for merge and clone
  - ziplock: The ziplock file created from ziplocking

chroma-externals: 15640/1

Change-Id: I6fcd8ce793431699954e49edce6d4457876a7082
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/15641
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Tested-by: Joe Grund <joe.grund@intel.com>