HYD-4641 Selenium: open_account_dialog does not wait for correct url
Jenkins pulls latest chrome automatically. We were on version 43 and now are on version 44.
Version 44 exposes a latent bug where we do not wait for the correct url when opening the
account dialog in open_account_dialog. We then make an async JS call that gets interrupted on page unload.

Make sure we wait for the correct url before proceeding.

  - Added a WebDriverWait call that waits until the url matches the user detail page.

Change-Id: Iabe94327ba626dda01d43ab5b275cbe2554f33ce
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/15696
Tested-by: Jenkins
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>