HYD-4662 Create d3 Line model
In our work towards generic charting, we need an easy way to create
lines.

Currently we use nvd3 for this, but it is buggy and not as configurable
as we'd like.

Instead of relying on this lib, we can create a simple
line model with very few lines of code.

This will enable us to use lines
easily and on demand (charts, spark-lines) and not be beholden to nvd3.

  - Added get-line module.
  - Added associated tests.
  - Added flushD3Transitions to
    move transitions to end states.

Change-Id: I98f83e600e18964d255a9fe723d3fccbb189bf8b
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/15805
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>