HYD-4678 Test failure: selenium, test_volumes, test_change_volume_config: element not visible
Couple of things to note for this bug:
1. It believes that new_primary_text and new_failover_text are the same.
If this were the case we would see a dialog with an "OK" button. I
cannot see the values based on the screenshot.
2. It cannot find the "OK" button because it is not in the modal that is
displaying. It should be trying to find the #transition_confirm_button.
3. I noticed the primary_text selection breaks from the loop when it
finds an item to select. The failover_text loop does not break. I added
in a break statement to be consistent.
4. I added some logging so the values can be compared.

Change-Id: If0439dde45454fa02dfd4dd4e0bd7bc0c94a9b86
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/15891
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>