HYD-4503 HYD-4690 HYD-4517 Support RHEL 6.7
Pacemaker in RHEL 6.7 has a serious segfault in it and that
Red Hat has elected not to fix because it only happens in
non-CMAN use patterns, which they don't support.

The solution is to build and/or use RHEL 6.6's pacemaker with 6.7.

Update the installer to allow EL 6.7 installations and
set the $TEST_DISTRIBUTION variable to test on EL 6.7.
Also need to update the nightly test runs to test on EL 6.6 and
EL 6.7 (Jenkins change).

To be able to test the 2.2->latest upgrade, hack the upgrade
to allow the installation of 2.2 on EL 6.7.

 - Update the chroma-agent.spec file to require the special
   pacemaker
 - Create a pacemaker directory in the chroma-dependencies
   to build our new pacemaker
 - Changes defaults to test on 6.7
 - Update the tests to allow install of previous version on any
   os version.

Change-Id: I1e5557be58234e0b50e6fb6bb68df084cde2c686
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17821
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>