HYD-4503 HYD-4690 HYD-4517 Support RHEL 6.7
Because the Pacemaker in RHEL 6.7 has a serious segfault in
it and Red Hat refuse to fix it because it only happens in
non-CMAN use patterns, which they don't support, we need to
build and/or use RHEL 6.6's pacemaker with 6.7.

Need to update the installer to allow EL 6.7 installations and
set the $TEST_DISTRIBUTION variable to test on EL 6.7.
Also need to update the nightly test runs to test on EL 6.6 and
EL 6.7 (Jenkins change).

To be able to test the 2.2->2.4.0.0 upgrade, hack the upgrade
to allow the installation of 2.2 on EL 6.7.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I1e5557be58234e0b50e6fb6bb68df084cde2c686
Reviewed-on: http://review.whamcloud.com/16174
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>