HYD-4695 Fix for false positives in EFS test
During the conversion of the EFS test to provisioner-ng
in HYD-4437, I fixed up a situation where false positive
(passes) of the test were being recorded due to not tracking
the return code of all iterations of the main loop.
I forgot to initialize a variable however causing a logic
error in that false positive hole closure.
Also, it's better to use -gt in [ ] than >.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I65ab2db77048d852a5f279e3abb7aa6040a35f23
Reviewed-on: http://review.whamcloud.com/15949
Tested-by: Jenkins
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>