HYD-4811 Add ability to set the LND Type with the NID
Until now IML presumed the ethernet is lnd type tcp and
infiniband was lnd type o2ib. However it is quite possible
and often required that people run tcp or o2ib over infiniband.
It is not possible to run o2ib over tcp.

IML needs to provide for allowing the use to configure valid
LND Types for network interfaces.

The approach is the the NetworkInterface endpoint will provide
a list of valid lnd types for that interface, this will allow
the GUI to provide a qualified list of types to the user. These
can then be programmed using the Nid endpoint.

 - Update Nid class and endpoint to support lnd_type
 - Update Nid class to provide a list of valid lnd_type for
   a given interface type.
 - Update NetworkInterface class and endpoint to provide
   the list of valid lnd_types for that class.
 - Added unit tests and integration tests.
 - Added validation to the nid and create a validation class
   that might allow us to easily validate other ends points.

Change-Id: I8c57d7a1acd874e52f8b0f38b9a5e72ba69eec3f
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15615
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>