HYD-4789 Simplify run_tests file
There is a repetitive pattern in the run_tests file that needs to be simplified. Once HYD-4788 is complete and all modules are using jasmine instead of minijasminenodewrap we can run all the tests within a for loop.

This patch will update the run_tests file such that repetitive code is encapsulated within a loop.

- package.json updated ui-modules dependency versions
- ziplock.json updated to match dependencies
- run_tests updated with for loop to simplify the commands used to run the tests
- updated chroma-externals link

Change-Id: I0e6f2d8cf034ae9f0ae5e9ca1dd5214bff7a8398
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/16272
Tested-by: Jenkins
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>