HYD-4775 Remove need for chroma-externals pragma
The chroma-externals pragma works well but has an overhead that
when a number of patches exist after a chroma-externals pragma
they must all have the commit message change.

We can change to use the sha1 in the original commit to find
the gerrit commit to pull.

Process is

See if git submodule update works
If not get the sha1 that we require from the error message, this
is not pretty but I couldn't work out how to do it another way.
query git to findout whic commit this sha1 is.
Fetch that commit from gerrit.
Finally to ensure all is well git submodule update should now work.

This patch also took the opportunity to create a standard shell process
run routine that can be called be anything and everything. This meant
providing a very thin layer in the agent to deal with agent requirements
and having some strangely capabilites in a 'standard' general routine
but did say replication and provide something that can be easily and
widely used.

 - New update_chroma-externals script.
 - Change calls to use the new script.
 - Changes to the agent to use the new agent layer.

Change-Id: I651caca7b56af02e39e711fa4d09d88502971e42
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17333
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>