HYD-4779 Add Robinhood 2.5.5-2
Add Robinhood 2.5.5-2 to the IEEL 2.4 Release.

This patch is actually a slight fiddle to get around an issue with
the build process. There does not seem to be a way to download and
build a 2.5.5-2 build. The make download works but the make rpms
gets very confused and tries to unpack 5.5-2-2. A ticket exists for
this but for this ticket the gz was renamed from 2.5.5-2 to 2.5.5
the output of build will be correctly labelled.

 - Download robinhood-2.5.5-2.tar.gz rename to robinhood-2.5.5.tar.gz
 - Update Make file for version and distro
 - Updated robinhood.spec file

Change-Id: I5e58174b4584a5f378d724922155b7d02cd2f394
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/16337
Tested-by: Jenkins
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>