HYD-4831 Upgrade tests might fail if the release tag is unchanged.
The upgrade test untars into the directory contained in the tar
file which is based on the release tag of the IML build. For
example when the last tag is IEEL 2.3.0 the directory will be
2.3.0/{components}, in the upgrade test if the upgrade is the same
tag then the upgrade will also untar into 2.3.0/{components}

This means that the 2.3.0 directory now contains a mixture of old
and new. The upgrade process will then by chance fill the repos with
either the new version or the old version or packages.

The version can be different for each bundle so

lustre server might be old
lustre client might be new
iml-agent might be old
etc.

Whilst the upgrade from the same tag could be avoided and the
old-directory could be deleted before the upgrade the best solution
is to place the old and new in different namespaces in the filesystem.

So unpack and run the old in

/tmp/previous_install

and the new in

/tmp/upgraded_install

This way all the information is available for debug and there is
no chance of name clash in the future.

 - Update installation_and_upgrade/run_tests to use directories
   as described above.

Change-Id: Ib26fa30bebbfc8dfc53f758d899dfc1a1617ef87
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/16415
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>