HYD-4867 Add CommandAlert to new alerting
Alerts used to be a hogwash of things alerts(many)+events(many)+commands.
Now they are just Alerts, although with all the flexibility of before.

However we lost the Command in the notifications and so need a CommandAlert
that starts when a command starts and finishes when it finishes.

Added 4 types of alert CommandRunningAlert, CommandCompleteAlert,
CommandErroredAlert and CommandCancelledAlert. This are created as the
command starts and then the alert morphs into one of the completed alerts

I.e. A fetchof id=X will produce a CommandRunningAlert until the command
completes then the same fetch of ix=X will produce a Command[Finished]Alert.

 - Create 4 new types of alert.
 - Add ability for the Alerts to be cast to different types.
 - Updates tests, although Abiola is writing additional tests for this.
 - Update the mocks appropriately.

Change-Id: I0179de532e20c4f8a18604888c6074387955dfd4
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/16669
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>