HYD-4879 Turn off Selenium tests on b2_3
The Selenium tests have long been in the process of being phased out and
replaced with other testing. This has actually completed entirely on the
3.0 branch. However, our b2_3 branch has a few remaining selenium tests
that are providing little value, but taking resources to track and
document and inhibit our ability to turnaround patches quickly (which
makes them a risk to timelines for releases off that branch). We believe
they are costing us more value than they are adding, so we are going to
remove the remaining tests from the b2_3 branch.

  - Making the bare minimum changes to get this done.

Change-Id: Ia8c2a20861cc2b23c0f236f0fdc46b8d3adfdf4a
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/16599
Tested-by: Jenkins
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>