HYD-4735 Need to include repos when updating
There is a defect where the iml, lustre and e2fsprogs repo is
not being enabled when determining the packages to be installed's
dependencies.

This patch tests a theory that we can with the existing codebase
update the agent on it's own, followed by the packages and hence
make existing systems upgradeable.

Just a patch for test.

Change-Id: I8a0d639a83cb030ef218707040bc8b68ba6fda52
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/16637
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
(cherry picked from commit 175587db4d658252e11f6fe61b7c618906f39c15)
Reviewed-on: http://review.whamcloud.com/16653