HYD-4951 De-duplicate provisioning code in IML tests
The code to call the provisioner and process its input/output 
is duplicated (inconsistently) across our various test scripts. 
This is difficult to maintain and a lot of tests miss out on extra functionality added to just some test scripts.

In order to de-duplicate the code and make it more consistent, 
this change pulls together the code that calls the provisioner 
and processes its input and output into one shared script 
that can then be invoked by all of the other tests scripts that need it.

  - Creates new scripts provision_cluster and release_cluster
  - Replaces provisioning-related code with call to above scripts

Change-Id: If9f2555e989217d454c36d5f1117e8236c2cda9e Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Change-Id: If9f2555e989217d454c36d5f1117e8236c2cda9e
Reviewed-on: http://review.whamcloud.com/16618
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>